Fix HTML tags in comments bug
This commit fixes the issue where DocumentFragment throws errors on parsing HTML tags inside HTML comments, and the editor freezes at the load screen. It now uses innerHTML to circumvent the problem, which is less performant but probably the best solution for now. As part of this, Ace editor has also been upgraded to version 1.4.14.